msm: kgsl: Add debugging for CP_INIT failures
Failure during CP_INIT (ringbuffer initalization) could be an
indication of any number of interesting and challenging issues.
Print a number of possibly useful register values to the kmsg in
addition to forcing a full snapshot.

Also reduce the timeout on any start operation from 20 seconds
down to 2 seconds. No init operation should ever take that long
and it allows us to debug issues quicker and gives the
rest of the system less time to pile on additional fences and
other things that might bring the world to a halt faster.

Change-Id: Ic0dedbad1ad5c5113670189f711a71170fa59acc
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>