msm: kgsl: Move the logic for the IOMMU sync workaround
As part of our effort to move more device specific configuration to
the device tree add qcom,quirk-iommu-sync to the device tree and
remove it from the device features.  We also know that the mutex must
be enforced globally so we can pull it out of the device specific
structure and streamline the usage slightly.

Change-Id: Ic0dedbadef6290dbc6d4e05372cd3014a4203e8a
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>